Added initializer for variable declaration. Refactored other AST node initializers.
AST node initializers now all follow the same pattern of declaring variables in the beginning, assigning values to these variables in the body and then assigning these variables to the properties (and calling super, if applicable) in the end, unwrapping the values in the variables if necessary. This allows many properties in AST nodes to become let's and non-optionals.