Implemented new AST parsing.
AST is now dynamic only (aka not type safe), but much more correct than the old Dynamic AST (since it correctly handles composed key names) and much easier to implement than the old GRYAst (it's mostly done). It also uses a lot less of the parser, so there's less code to maintain (and to be error-prone).