Added a method to seed from the last hour instead of the current hour.
If a bug is found with the random number generator and the hour passes, this method allows easy changing of the seed for repeating the bug for another hour. All that's needed is to change the `rng` variable (immediately below the method) to seed with the `seedFromLastHour` method.

If another hour passes, or if the user wants to record that specific seed, the method is easy to understand and change accordingly. All used numbers have very intuitive values. Specifically, weekdays are counted starting in Sunday (meaning Thursday is day `5`) and hours are counted in 24h style (so 2PM is `14`).