Removing Xcode project from repo
Swift Package Manager can automatically generate and re-generate Xcode Projects (this one was generated by it, for instance), so it doesn't need to be in the repo.

SwiftPM also makes it easy to build, run and test the package from the command line, editing it with another IDE or text editor (other than Xcode). If someone were to use another editor, especially in Linux (since Xcode doesn't even exist there), the xcodeproject in the repo would be more than useless.

Finally, the SwiftPM's automatically generated .gitignore adds a line to ignore all Xcode Projects, indicating an official opinion that they shouldn't be in the repos. This one was here only because it was created before the migration to SwiftPM (and before the creation of the .gitignore), and therefore it had to be removed manually with `git reset name_of_file` and `git rm --cached name_of_file`.