Improved fetching of the test directory
Now the program must always be run from the package's root (it will find the test files from there) instead of the user passing the file's location via command line.

This is needed because it seems I can't pass the test directory when running from the actual command line (neither with `swift test` neither with `./.build/.../GryphonPackageTests.xctest`). Any arguments passed this way are interpreted as "execute only this test file", not as "pass this as an argument to the executable". The `swift test --help` page didn't provide any information.

This is fine for my workflow: running from Xcode there's an option to set the working directory to `$(SRCROOT)`, and running in Docker for Linux I always run from the package's root. It might not be fine for potential future contributors, but it can be adapted later if necessary.