Added initial (extra) tests to GRYSExpressionParser.
These tests compare the AST generated dynamically by the parser (which might change and break over time) with an AST generated at the time of the writing of the tests (which presumably is correct, but is not thoroughly checked or tested).

This is meant to be like a regression test: any changes to the parser that affect these ASTs will be reflected instantly in the test suite. The test suite can then be updated, if those changes are harmless.