Translate print as println
Treats the `print` function as a special case, for now. Whenever it's called we translate it into `println` instead of `print` to obtain a similar behaviour in kotlin.

This is being done only because the print function can help in the decelopment of integration tests, and it's good to have those early on to aid in the development of other features.

This replacement (or something like it) will eventually be made by a more sophisticated system meant to process standard library functions, types, etc and then translate them appropriately.