Improves onboarding of new contributors
Adde the Dockerfile to the repo so anyone can test in the Linux environment (now that the tests use the kotlin compiler, it must be installed on docker, so the Dockerfile is no longer trivial).

Improved the Readme to give details on contributing and setting up the environment.

Added the latest changes to main.swift, so people have an better understanding on how to use the API.