Implements initial version of kotlin literals.
The translation now counts with a kotlin literal function GRYKotlinLiteral. When called from swift, it returns exactly the swift value passed to it (and will probably end up being optimized away). When translated into kotlin, it disappears and gets replaced by the kotlin string passed to it. This way, we have a manner of "conditional compilation" that's easy to use and to implement.

This commit includes the implementation and the tests for the new feature. It also includes a bugfix: strings containing escaped double quotes are now parsed properly.