Adds support for doubles created from integer literals.
Kotlin Doubles can't be initialized from integer literals, they must be initialized from double literals. With this change, when the translator finds a Double being initialized from an Integer literal, it adds  ".0" to the literal to transform it into a Double literal in kotlin.

This might seem superfluous (since we could just always use double literals in swift) but it will be needed later on for floats (kotlin's 0.0f, which swift doesn't support).