Bugfix: removes extra newlines in some Kotlin literals.
In a previous bugfix, call expressions stopped having a newline at the end (so that we could have inline function calls); the responsibility of adding that newline went to the statements.

That bugfix revealed another bug: kotlin literals added an extra newline to compensate for it. That newline has now been removed.