Compilers now load json before generating Kotlin.
This speeds up significantly compiler runs on files that have up-to-date jsons. That's because it loads the cached AST from JSON and therefore skips generating the AST from the .ast file, which is by far the slowest part of the process.

This won't speed up tests but will significantly speed up some "main" runs.