Moved most of the horizontal limit logic to Printable As Tree.
This is the ideal place for this logic. It's now accessible to all printable tree instances, it's no longer dependant on global state, and it can be tested.

Also, it might be more performante than both previous alternatives. It doesn't need to create an array like the second version, and it only calls the `printFunction` (which can be costly) once, which probably offsets the cost of an extra String instantiation.