Corrected the translation of if expressions.
If expressions wrap their real conditions in an "empty" call expression. I previously thought that empty expression was part of the boolean literal, but it seems to be part of the if expression itself. This is evident when we try to translate an if using an explicit function call as a condition (`if someFunction() { ... }`). The code now reflects that fact and allows proper translation of function calls as conditions.