Fixed sorting for Printable ASTs.
Sorting was added here to guarantee that an AST was always printed the same way. This is needed for the dictionary (since iterations in dictionaries with the same elements may result in different orders) but not for the array, since its elements are always added in the same order and therefore are guaranteed to always be iterated in that order.