Added local test files to the gitignore.
The test.swift file in the root of the package is used to quickly test changes to the compiler. Therefore, it's constantly changing, but it's contents aren't at all supposed to be in the repository, and tracking its changes would be rather useless.

The test.swift file itself was already being ignored, as was the generated test.ast file. However, the (also generated) test.json file was not, and therefore it was always showing up in the commits (annoyingly so). Now, the gitignore ignores all test.* files, as it should have done from the beginning.