Separated json files for tests and for caching.
Json files were being used for two things: for the caching and for testing the parser.

The caching part involved several tasks reading the json files directly instead of processing the ast dump, since this skipped the parser and resulted in a significant speedup for those tasks.

The parser tests consisted in parsing the AST and comparing it to an old version of the json known to be right. However, since the same json files were being used for caching they got updated before the tests could run. Therefore, the tests were comparing a new json with another new json, since the old (and known to be correct) json was already lost.

Now, caching will use *.json files and parser tests will use *.expectedJson files. In order to generate these expectedJson files, the main file was updated with new functions for this. The parser tests should always run using old expectedJson files; these files should be updated _after_ the parser tests have run, and the program is known to be correct (i.e. only and always before any commits).