Added back the parser tests.
The parser tests were hanging before because the ifStatement test case was outdated, causing the test to fail, causing the diff tool to be called - which was hanging. Updating the json file fixed the test so that it passes, but the diff still needs to be fixed so that it doesn't hang in future failures.