Rewrites the parser to use manual indices.
With this rewrite, the parser now manually counts its indices instead of allocating Strings and Substrings everywhere. Consequently, the parser's buffer is never changed. This means the code is harder to read, but it also means parsing the ifStatements test case goes from an average of around 11 seconds to 0.5 seconds.

Some of the rewriting still needs cleaning up, but since this version is passing all tests I'm committing it.