Fixes val/var declarations in computed properties.
The val-var keyword in kotlin works a bit differently than the let-var keyword in swift. In kotlin, it indicates that the property doesn't have a setter, while in swift it indicates that the property will never change.

This is important because (for instance) computed properties with getters that return different things at different times can be val's in kotlin (because they can have setters) but they must be let's in swift.

The translator now bases its decision on using val vs. var not only on the swift choice of let vs. var, but also on the presence or absence of getters and setters.

This allows us to finally translate a (heavy modified) version of the Calculator app! Woo-hoo!