Adds the GRYIgnore protocol to ignore classes and enums.
Classes and Enums that conform to the GRYIgnore protocol will now be ignored by the translator and won't show up in Kotlin.

Also fixes a bug in the parser where it would crash when inheritance clauses were the last item in a subtree (as in the declaration of an empty enum conforming to GRYIgnore).