Adds implementation and tests for arrays.
Finishes the WIP implementation of arrays, including subscript getters and setters, array literals, and the ArrayReference type.

ArrayReference is new a class that wraps a swift array, making it pass-by-reference instead of COW. This is the same behavior as Kotlin's arrays, meaning that using this new type in Swift allows both the original code and the generated code to work the same way.

To use this type, the user just has to add an explicit `ArrayReference` type annotation (see `Test Files/arrays.swift`). The type conforms to the same protocols as Swift's native Array implementation, which means their API's should be the same.

At this point, performance has not been seriously considered (neither for Swift nor for Kotlin.