[GRYAst] Started adding a new AST layer
Newer versions of the compiler use different AST layouts, as the AST is essentially only an internal representation. In order to deal with this issue (and start supporting Swift 4.2), a new layer is being added between the Swift AST and the translation to Kotlin.

The old Swift AST, named GRYAst, is now called GRYSwiftAST. It is meant to be a generic data structure, capable of representing the AST of any version of Swift.

The GRYSwiftAST will be translated into a new GRYAst, which will be Swift-version-independent. This translation will be done by version-specific translators like the new GRYSwift4_1Translator. These translators should encapsulate all the version-specific knowledge on how to turn a Swift AST into a Gryphon AST.

The GRYAst will then be translated into Kotlin code by the GRYKotlinTranslator, now much lighter than before.

It also seems (for now) that the Swift Translators will contain most of the logic that was previously on the GRYKotlinTranslator, making it much slimmer and easier to understand. It also seems that the GRYAst, being an internal representation, will be well-suited for layers of optimizations and corrections.

This commit also includes a few leftover improvements to diagnostics. These changes, however, aren't taking into consideration diagnostics much, as they'll probably have to be completely re-thought after this is done.