Cleanup
Test files are now updated when the tests are run instead of the main program. This is good because the main program hasn't needed to update its files since the S-ExpressionParser became fast enough to make AST caching unecessary, but the test cases do need their files updated. Now the files get updated if and only if it's needed.

Also improves readme descriptions and adds some TODO's.