Adds pass to remove extraneous parentheses where it's safe.
An unintended (and unfortunate) consequence of this approach is that string literals inside escaped parentheses in interpolated strings get interpreted as part of the normal string and get "mistranslated" (see the changes in print.swift). However, this should hardly happen in practice and the resulting translation should still behave correctly, so it's an acceptable problem for now.