Starts to add throws to Kotlin translator for expected pass results.
The GRYAst data structure is the same before transpilation passes and after. This means that certain invariants can't be coded into the type system. For instance, `extensionDeclarations` are expected to be present before passes, but should have been completely removed by the `GRYRemoveExtensionsTranspilationPass` afterwards. However, as they're coded into the GRYAst type, the compiler can't guarantee their absence after the passes, and the Kotlin translator has to check for them anyway at runtime.

By throwing errors in the Kotlin translator when extensions are found we can document in the code these types of expected preconditions, even if they can only be checked at runtime.