Cleanup - Unify file coding API
The API for encoding into a file and decoding from a file is now unified into GRYCodable protocol extensions. The only thing that remains separate is decoding a GRYSwiftAST from a Swift AST dump, as that is a separate operation that doesn't follow the common Codable processes despite using some of the same logic and code.