Bugfix - Fix probabilistic test.
Since the Codable test is done by comparing types and descriptions of values (instead of the values itself), it could happen that dictionaries would generate different descriptions of themselves (i.e. with different orders for entries, since that order is probabilistic). Now the test should be deterministic, though a bit weaker.