Feature: Bootstrap - Add support for Open Existential Expression.
It's not clear to me at the moment exactly how these expressions work, but for the two examples I've seen so far, this algorithm seems to work.

As this expression isn't clear and as it is, in a way, an implementation detail, its test will be just the bootstrapping of the GRYPrintableAsTree file (where it was found), at least for now.