Feature: Bootstrap - Start adding support for Swift 5
Swift 5 support is needed as it can output the AST of multiple files at the same time, thus enabling us to get the ASTs for the whole transpiler (as the files have to be compiled together).

This commit adds the Swift 5 Translator class, (still untested) support for String literal interpolations, as well as a few other minor fixes.

It also adds an output file map to enable dumping the AST for the transpiler in the future.