Bugfix: the previous implementation broke Swift 4.
Swift 4 negative numbers contain both a wrapping unary expression and a "negative" attribute on the literal expression, so this method was leaving the literals with two negative signs (one for the unary expression and one for the negative number). Now we check for one or the other, depending on the Swift version.