Test - Add tests for ArrayReference
Adds some tests for the manual implementations of ArrayReference. Also adds two `appending` methods, to be used as mutating versions of `append` instead of the `+` operators from before. For some reason, Swift 5 typechecks a sum of two `ArrayReference`s ignoring this manual definition and using instead a `+` operator from `RangeReplaceableCollection`, one that accidentally modifies the left-hand-side array, making it dangerous to use. Having an alternative in the `appending` methods might mitigate this issue, though it's not really a fix.