Refactor - Unite code for the standard libraries in a single file
The new file contains:
- Code for literal manipulations in Swift, like `GRYAnnotations`, which shouldn't really be declared in individual files like GRYPrintableAsTree.
- Classes and other declarations that should directly affect Swift code, such as ArrayReference.
- Classes and other declarations that should directly affect Kotlin code, such as `MutableList.copy()`.

Keeping all this in a single file is good for separating this responsability from the rest of the code, as well as keeping track of the growth of the standard library so far.

It might be inconvenient to keep Swift-only code in a file that gets transpiled. If this becomes more of an issue, it might be interesting to separate the file into one for the Swift stdlib and one for the Kotlin stdlib.