Feature: Error Handling - Allows compiler to continue after errors.
Most functions that threw errors now only throw when the compiler is set "stop at first error". Otherwise, they log the error and insert new `error` nodes in the AST, which will then be largely ignored by the subsequent passes. This should allow us to have a better idea of how many errors happen in a given translation, and also where they are (as a full translation replaces error nodes with  "<<Error>>" in the output code, which can be a useful visual guide to finding the source of the problem).

It's worth noting that the GRYDecoder does not use this new error handling method, as it currently can't recover well from errors when it finds them.