Feature: Swift coverage - Add support for dicionary literals
Also fixes a few issues with closures and tuples:
  - Removes explicit types from closures, which might add some flexibility to dictionary method translations.
  - Implemented translations of tuple elements, which also lends some flexibility to dictionary method translations. It might lead to misleading problems due to incorrect tuple translations (as tuples themselves don't exist in Kotlin) but I think that these problems would be easy to identify and fix.
  - Adds a small fix where translating types would remove the parentheses around tuple types as if they were redundant. It might seem weird, since the translated tuple types won't compile in Kotlin anyway, but it might make the resulting (incorrect) Kotlin code easier to understand and diagnose.

This also removes the dictionary fix from Swift 4, which technically means that we don't support Swift 4 anymore. This should be fine as Gryphon's still in beta (no one is depending on Swift 4 support) and Swift 4 can't handle dumping multiple ASTs in any case. However, it might be better to unify the two Swift translators now.