Refactor - Start refactoring the AST to use structs instead of enums
AST nodes were represented as enums with associated values, but some of them contained a lot of information, which made unpacking and repacking those enums a very verbose part of the code. Also, the syntax of this unpacking and repacking made it hard to know which properties of the node had been changed in certain places, something that became evident in this refactoring.

By switching the associated values with mutable structs, we can just use or change the parts that are interesting for us at each point, thus (apparently and hopefully) making the code easier to understand.