Feature: Correctness - Translate functions using implementation labels
The way functions were translated before was easy to implement but caused a problem for their implementations. Functions with different api labels (i.e. `f(a b: Int)`) ended up using api labels as the variable names, which caused weird variables names like `from` and could potentially cause errors (if two or more parameters had the same api labels).

Now, functions that are declared in a file interpreted by Gryphon get registered so that all references to them use their internal argument names instead of the api labels. This allows the implementations to use the internal names as well, avoiding the ambiguity problem. It also (I believe) creates more kotlin-y code.

To do that, it was necessary to add support for tuple shuffle expressions (in order to deal with shuffled tuples in function calls involving variadics and default arguments).

Also fixes an issue with the closures.swift test case (the test case was incorrect and didn't compile successfully).