Feature: standard library - Initial support for comments as annotations
This adds initial support for reading comments from the source file. Comments can only be read if they come after AST nodes that include a `range` attribute.

In this commit there's support for reading comments as annotations for enum cases. Enum cases can't have expressions, so they couldn't use `GRYAnnotations` like the others. This should allow users to add annotations to more places than before.

In time, it would be interesting if most (if not all) of the special code manipulation mechanisms were turned into comments.

The main thing missing here is a way to get the source file and pass it to `GRYSwift4Translator`. It might be a good idea to make the translator fetch the source file from the name it gets in the start of the AST dump.