Feature: standard library - GRYAlternatives are now comments
This is a more limited solution because it does not allow replacing values inside an expression (only replacing the root expressions). For instance, what was done in kotlinLiterals.swift around line 42 is no longer possible. However, it is believed that these cases can be refactored into separate variables, or styled into multiple lines (like function declarations and calls) each with their own (optional) comment.

The only comments considered by this implementation are the ones at the last line of the expression. All others should be ignored.