Refactor - Remove the GRY prefix
Removes the GRY prefix from most of the places it appears in the code. This should make it a bit easier to maintain (less mental overhead for reading and understanding). Disambiguation can be done by namespacing. If this eventually becomes a problem, we can think about revisiting this decision.

The GRYCodable types haven't been changed yet because it's likely they'll be removed soon.