Bugfix - Remove placeholders from bootstrap's AST dumps
The placeholders were initially meant to avoid hard coded paths when committing AST dumps. Since these files are never committed, we can remove them. This solves an issue where the script would replace the swift file's path with a placeholder and then replace the placeholder with the path to the AST dump (with its extension changed to .swift), even though the Swift file and the AST dump file could be in different locations.