Refactor - Remove caches and GRYCodable
The idea for the KotlinTranslator, SwiftTranslator and TranspilationPass tests (as well as some of the GRYCodable tests) started out when the transpiler was still small. In that situation, it was feasible to manually check if the changes in the cache files made sense before updating them. For a while now, however, changes to the cache files were being ignored, and just checking the output and using lldb to check the data structures in the middle was solving the problem.

Another reason to have these caches was the fact that early on the ASTDump parser was really slow, but now the time it takes is negligible.

With these two changes, it now doesn't make much sense to keep these files around. This means that the  GRYEncoder could be deleted, as well as several test cases that only used this method, and many other things that became dead code along the way.

This commit also deletes dead code resulting from removing the fuzzer for String.split (as the method got more powerful, maintaining a fuzzer for it wasn't worth it, and the fuzzer never did catch any errors). This includes most of the custom RNG logic, as well as the code from Cocoa With Love.

With these changes, the transpiler becomes a lot simpler and easier to maintain and develop. The removed tests were being ignored, so there's no loss there.

It might be interesting to create new tests that actually work for these classes. For instance, manually hard-coding a SwiftAST in a test, translating it, and checking it against a manually hard-coded GryphonAST.