Refactor - move the responsibility to create a SwiftAST to the decoder
The method for creating a Swift AST has a lot of knowledge about the internal workings of the Swift AST dump and the decoder, and no knowledge about the inner workings of the SwiftAST itself. Moving the test to the decoder makes makes the code's contents clearer.

Also removes the placeholders for file paths and memory addresses. These were used in the beginning of Gryphon's development to enable swiftASTDump files to be committed, but since they're no longer on the repo it's not needed anymore.