Bugfix - Revert capitalizing unknown enums
Some static method calls are represented in the AST in the same way as enum cases. Because of this, the previous method of capitalizing all enum subtrees doesn't really work (as it would also capitalize static methods). The pass now only capitalizes known enums, and unknown enums will just have to manually changed to have the same name in both languages.