Standard library: translate description variabels as toString functions
The description variable in Swift has generally the same behavior as the toString function in Kotlin: it's the main entry point for getting an object's description as a string to be printed. This translation allows for a more adequate behavior in the translated code.

It might be better for this translation to be with the standard library templates, but that's not possible with the current system. If this type of translation comes up more often, however, it might be useful to create such a system.