Refactor - Add support for user-defined templates in code
Users can now define a gryphonTemplates() function in their code that will be treated by the compiler like a library template file. This should allow users to specify translations for their own files, like (for instance) translations of functions that have to be declared separately (i.e. the `occurrences(of:)` function in Extensions.swift).

Also organizes the passes in TranspilationPass to try to make them easier to understand.