Bootstrap - Refactor occurrences to be translatable
The occurrences function before was being ignored and replaced by a manual translation. This changes the function's implementation to be almost the same for both languages (except for a couple of sticky points in dealing with the standard library and with substring indices).

The code now looks maybe less nice, but is has the added advantage that all changes that may come up in the Kotlin-specific lines will be next to their Swift counterparts, which hopefully makes them harder to miss.

Also adds support for tuples in SwiftTranslator to enable translating a Range.init method that uses tuples.