Release - Print warnings in the Xcode format
This allows warnings (like the ones thrown by the Transpilation Passes) to be printed in the same Xcode format as the SwiftTranslator errors. This format includes printing the source file line that caused the problem, as well as the line number and column.

For this to work, the GryphonAST now includes a reference to its SourceFile, which enables TranspilationPasses to access it.

Also, the logic of rendering the Xcode format was moved to the Compiler so it'd be accessible to all parts of the compiler.