Ease of use - Automatically create rawValue variables in enums
Enums that conform to rawRepresentable have their conformances automatically created. The Swift compiler includes the AST for that variable in the AST dump, but unfortunately it's not well formed by the time of the AST dump. A quick investigation of the Swift source code indicates that this is not a problem with the dumping itself, so it wouldn't be that easy to fix.

Instead, the transpiler now creates the rawValue variable by hand.