Bootstrap - Translate Utilities top level variables
These variables must be translated in top-level because static vars in extensions aren't allowed in Kotlin.

In order to do this, we need to teach the compiler not to treat any file as a potential main file, otherwise these variables get sent to a main function.

Now, compilations of multiple files are assumed not to have a main file, and compilations of only one file are assumed to only have a main file.