Release - Implement driver
Implements a Driver to handle interpreting command line arguments and running the Compiler steps. This should allow the program to be run from the command line normally.

It should also allow the bootstrapped version to be easily called (without being recompiled) in order to make it emit several compiler steps that can then be tested against the Swift version's results. For instance, the bootstrapped version's SwiftAST has to be compared to the original's to guarantee that the AST dumps are being interpreted correctly.