Test - Refactor Bootstrap test's setup
Sent most of the setup to a separate bash script that must be called before running any of the Bootstrap tests. Also changed the BootstrapTest.setup method to only detect if the script needs to be run again or not.

This should be simpler, easier to understand, faster, and do less repeated work.