[WIP] Add support for calling superclass's initializers
Superclass initializers in Kotlin have to be called at the header of the function declaration, rather than in the body. This pass refactors those calls into the header, which should work fine for simple cases (such as the ones introduced by this Statements/Expressions refactoring).

In order to do that, we also added an InitializerDeclaration class and removed the FunctionDeclarationData class, two things that were probably going to be done anyway but have made this process simpler.

This commit also makes InitializerDeclaration be used in other places that check for FunctionDeclarations whose prefixes are "init". This probably wasn't necessary now, but it's one less thing we have to do later.