[WIP] Remove duplicate declarations from rearranged if-lets
Rearranging if lets that had the same declarations in different else-ifs can cause those declarations to be duplicated in the resulting Kotlin code. This should remove the duplicates, hopefully without creating correctness issues in the process. The only problem that's foreseen right now is if removing one of the declarations causes some side-effects to not happen, but this should be taken care of by the warnings from before.