[WIP] Bugfix
This was probably caused by the same Swift version change that motivated some fixes in SwiftTranslator a couple of commits ago. This caused Member Reference Expressions to be ignored (they were translated only as the left side) because the right side is in the member expression itself, which was getting skipped by the `subSubExpression` bit there.

That said, this project really needs better test coverage (especially in the Test Files folder) to ensure that bugs like these are detected earlier and the fixes don't cause issues in the translations that motivated this code in the first place.