Refactor - Update a few of TranspilerPass's methods' arguments
The methods in TranspilerPass used to receive many arguments to avoid having to de-structure the enums from the AST repeatedly. However the AST now uses classes, and we need these methods to also receive information on their ranges. Rather than adding ranges as parameters to all methods, we're re-structuring the parameters to only send in the classes as they are.

This will probably also help in the future when we try to refactor the TranspilationPass to mutate the AST rather than replace it.