Refactor - Initial implementation of a GryphonType type
This implementation is an attempt to replace the current system of dealing with types as Strings. It should be more correct and easier to understand.

So far, it gives all the same answers for subType relations as the current algorithm when run with the unit test suite. Bootstrap tests weren't attempted yet.

Initial tests also show a significant slowdown. This may be because of overhead from the extra debugging tasks that were performed, but it may also be from all the extra ArrayClass (and maybe String) allocations. An interesting test might be to try to create small structs capable of holding three or four Strings and using them instead of arrays to make most cases faster.