Refactor: Extract KotlinTranslator's state into a TranspilationContext
Several classes were accessing static state in KotlinTranslator, which broke modularity but more importantly could cause issues if the program tried to transpile more than one program in a single run. This is a first step towards making this state an instance that will be passed on to each transpilation step to solve these issues.