Feature: multiline string support
Also refactors the stdlib contents in the Driver to make better use of multiline strings, and sends it to another file so as not to crowd the Driver file with unrelated content.

The explicit `kotlinStringInterpolation` avoids issues with the `${_string}` being interpreted as a string interpolation in Gryphon's Kotlin codebase, as opposed to it jut being literally written in the String. Perhaps future versions could escape these patterns during transpilation.