Feature: Automatic creation and calling of AST dump scripts for Xcode
The `xcodebuild` command line tool allows us to easily detect the project's call to the Swift compiler, which lets us automatically create a script for the AST dump and call it from Gryphon.
This is a first step toward making it easy to use Gryphon with real iOS apps using Xcode. It is also the start of a focus on using Gryphon with Xcode instead of a standalone tool, which will probably end with some refactoring/rethinking of existing flags and commands.

- Also prints an error whenever Gryphon fails to write to an output file (i.e. when that file is in a folder that doesn't exist yet). Before that operation used to fail silently and the user might think it had been completed successfully.