Feature: Don't change current directory for transpiled executables
Before, specifying -run meant Gryphon would compile the transpiled files into  a `kotlin.jar` file in a certain directory and then change into that directory to run the executable. This change makes it so that the directory isn't changed -- that is, the executable is run from the current directory. This should allow Bootstrap tests (compiled to `Bootstrap/kotlin.jar`) to run from the root Gryphon directory instead of the Bootstrap directory, which is necessary for the UtilitiesTest.