Refactor: kotlin unit tests now report failure explicitly.
Kotlin unit tests used to count the number of passed tests and fail if any other messages were displayed. This was a problem because some tests, like DriverTest, print error messages as a side effect that should not be considered. To avoid this issue, the tests now search for explicit messages of test failures.