Test: Improve test reporting for Kotlin unit tests
As more tests get added to Kotlin, so does the extra information they print. These changes make the test results be printed at the end of the test run, so that the results are easier to see. It also has the added advantage that tests no longer need to remember to call `super.runAllTests()` at the end.