Bugfix: change appending type to match Kotlin
In Kotlin, many MutableList methods return a List instead of a MutableList. Since the conversion to FixedArray is cheap, it might be better to match that behavior in other methods too. This would allow users to better reason about their Swift and Kotlin codes, and would also reduce the complexity of having to add calls to `toMutableList()` in translations.