Feature: Improve Map and MutableMap support
Also, add translations for tuples as Pair's. This causes an issue with methods on dictionaries, since key-value pairs need to be translated as Entry<Key, Value> instead of Pair<Key, Value>. The workaround for that (temporarily, I hope) is to translate any tuples with "key" and "value" names (i.e. "(key: String, value: Int)") as dictionary entries. This will probably break some things later on (for instance, tuples with this name that aren't related to dictionaries).
- Also, removes the reduce test that was causing issues with this new workaround.

Note: These latest changes (into using Lists and Maps) actually cause a small slowdown on the second batch of TranspilationPasses, probably because of the new MutableLists that have to be manually generated out of normal Lists. This will probably be solved once TranspilationPass'es are refactored to use in-place transformations.