Revert "Cleanup: add a local template to the global templates file"
This reverts commit 451eb1fa4dc178e6ccdddcb44a620824de34795f.

The template was in this file because it required a translation added in KotlinStandardLibrary.kt - meaning it can't be in the global templates file unless we demand that every translation should link the library files in order to compile successfully. Using an expression that triggers this template would create a translation that references a method that only exists in the KotlinStandardLibrary, and without it the error message is simply that the method was not found, which is less than helpful.

This commit adds comments to explain this dynamic and make sure a mistake like the reversed commit doesn't happen again.