[WIP] Feature: flag to avoid using Unicode characters
Unicode characters were causing problems when running the Bootstrapped compiler in Docker. Kotlin in Docker seems to replace some characters (the accent in "Vinícius", the box and arrow characters in tree printing) as "?". This was causing the bootstrap test to fail in Docker, since the bootstrapped compiler's output didn't match the Swift version. This commit adds a flag that replaces those characters with others, and removes the accent from my name in comments, which makes the bootstrap tests work in Docker.

Also: improve comments in PrintableAsTree.